move utils functions to related modules
The functions in `strain_mappings_generator.py` were moved to relevant modules, such as `genomics/utils` and `metabolomics/utils`
