+++ src/nplinker/arranger.py
@@ -6,19 +6,19 @@ from pathlib import Path
 from jsonschema import validate
 import nplinker.globals as globals
 from nplinker.config import config
-from nplinker.genomics import generate_mappings_genome_id_bgc_id
+from nplinker.genomics.antismash import podp_download_and_extract_antismash_data
+from nplinker.genomics.bigscape.runbigscape import run_bigscape
 from nplinker.genomics.mibig import download_and_extract_mibig_metadata
+from nplinker.genomics.utils import generate_mappings_genome_id_bgc_id
 from nplinker.globals import GENOME_BGC_MAPPINGS_FILENAME
 from nplinker.globals import GENOME_STATUS_FILENAME
 from nplinker.globals import STRAIN_MAPPINGS_FILENAME
 from nplinker.metabolomics.gnps import GNPSDownloader
 from nplinker.metabolomics.gnps import GNPSExtractor
-from nplinker.strain import podp_download_and_extract_antismash_data
-from nplinker.genomics.bigscape.runbigscape import run_bigscape
-from nplinker.strain.strain_mappings_generator import podp_generate_strain_mappings
 from nplinker.schemas import STRAIN_MAPPINGS_SCHEMA
 from nplinker.schemas import USER_STRAINS_SCHEMA
 from nplinker.schemas import validate_podp_json
+from nplinker.strain.utils import podp_generate_strain_mappings
 from nplinker.utils import download_url
 from nplinker.utils import list_dirs
 from nplinker.utils import list_files
