+++ src/nplinker/genomics/__init__.py
@@ -1,21 +1,11 @@
 import logging
-from .abc import BGCLoaderBase
 from .bgc import BGC
 from .gcf import GCF
-from .utils import add_bgc_to_gcf
-from .utils import add_strain_to_bgc
-from .utils import generate_mappings_genome_id_bgc_id
-from .utils import get_mibig_from_gcf
 
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
 __all__ = [
-    "BGCLoaderBase",
     "BGC",
     "GCF",
-    "add_bgc_to_gcf",
-    "add_strain_to_bgc",
-    "generate_mappings_genome_id_bgc_id",
-    "get_mibig_from_gcf",
 ]
