+++ src/nplinker/genomics/utils.py
@@ -6,9 +6,12 @@ from jsonschema import validate
 from nplinker.globals import GENOME_BGC_MAPPINGS_FILENAME
 from nplinker.logconfig import LogConfig
 from nplinker.schemas import GENOME_BGC_MAPPINGS_SCHEMA
-from nplinker.strain_collection import StrainCollection
+from nplinker.schemas import validate_podp_json
+from nplinker.strain import StrainCollection
 from nplinker.utils import list_dirs
 from nplinker.utils import list_files
+from ..genomics.antismash.podp_antismash_downloader import GenomeStatus
+from ..genomics.antismash.podp_antismash_downloader import get_best_available_genome_id
 from .bgc import BGC
 from .gcf import GCF
 
@@ -179,3 +182,128 @@ def get_mibig_from_gcf(gcfs: list[GCF]) -> tuple[list[BGC], StrainCollection]:
                 if bgc.strain is not None:
                     mibig_strains_in_use.add(bgc.strain)
     return mibig_bgcs_in_use, mibig_strains_in_use
+
+
+# ------------------------------------------------------------------------------
+# Functions to extract mappings for genomics side:
+# strain_id <-> original_geonme_id <-> resolved_genome_id <-> bgc_id
+# ------------------------------------------------------------------------------
+def extract_mappings_strain_id_original_genome_id(
+    podp_project_json_file: str | PathLike
+) -> dict[str, set[str]]:
+    """Extract mappings "strain id <-> original genome id".
+
+    Args:
+        podp_project_json_file: The path to the PODP project
+            JSON file.
+
+    Returns:
+        Key is strain id and value is a set of original genome ids.
+
+    Notes:
+        The `podp_project_json_file` is the project JSON file downloaded from
+        PODP platform. For example, for project MSV000079284, its json file is
+        https://pairedomicsdata.bioinformatics.nl/api/projects/4b29ddc3-26d0-40d7-80c5-44fb6631dbf9.4.
+    """
+    mappings_dict = {}
+    with open(podp_project_json_file, "r") as f:
+        json_data = json.load(f)
+
+    validate_podp_json(json_data)
+
+    for record in json_data["genomes"]:
+        strain_id = record["genome_label"]
+        genome_id = get_best_available_genome_id(record["genome_ID"])
+        if genome_id is None:
+            logger.warning("Failed to extract genome ID from genome with label %s", strain_id)
+            continue
+        if strain_id in mappings_dict:
+            mappings_dict[strain_id].add(genome_id)
+        else:
+            mappings_dict[strain_id] = {genome_id}
+    return mappings_dict
+
+
+def extract_mappings_original_genome_id_resolved_genome_id(
+    genome_status_json_file: str | PathLike
+) -> dict[str, str]:
+    """Extract mappings "original_genome_id <-> resolved_genome_id".
+
+    Args:
+        genome_status_json_file: The path to the genome status
+            JSON file.
+
+    Returns:
+        Key is original genome id and value is resolved genome id.
+
+    Notes:
+        The `genome_status_json_file` is usually generated by the
+        `podp_download_and_extract_antismash_data` function with
+        a default file name defined in `nplinker.globals.GENOME_STATUS_FILENAME`.
+    """
+    gs_mappings_dict = GenomeStatus.read_json(genome_status_json_file)
+    return {gs.original_id: gs.resolved_refseq_id for gs in gs_mappings_dict.values()}
+
+
+def extract_mappings_resolved_genome_id_bgc_id(
+    genome_bgc_mappings_file: str | PathLike
+) -> dict[str, set[str]]:
+    """Extract mappings "resolved_genome_id <-> bgc_id".
+
+    Args:
+        genome_bgc_mappings_file: The path to the genome BGC
+            mappings JSON file.
+
+    Returns:
+        Key is resolved genome id and value is a set of BGC ids.
+
+    Notes:
+        The `genome_bgc_mappings_file` is usually generated by the
+        `generate_mappings_genome_id_bgc_id` function with a default file name
+        defined in `nplinker.globals.GENOME_BGC_MAPPINGS_FILENAME`.
+    """
+    with open(genome_bgc_mappings_file, "r") as f:
+        json_data = json.load(f)
+
+    # validate the JSON data
+    validate(json_data, GENOME_BGC_MAPPINGS_SCHEMA)
+
+    return {mapping["genome_ID"]: set(mapping["BGC_ID"]) for mapping in json_data["mappings"]}
+
+
+def get_mappings_strain_id_bgc_id(
+    mappings_strain_id_original_genome_id: dict[str, set[str]],
+    mappings_original_genome_id_resolved_genome_id: dict[str, str],
+    mappings_resolved_genome_id_bgc_id: dict[str, set[str]],
+) -> dict[str, set[str]]:
+    """Get mappings "strain_id <-> bgc_id".
+
+    Args:
+        mappings_strain_id_original_genome_id: Mappings
+            "strain_id <-> original_genome_id".
+        mappings_original_genome_id_resolved_genome_id: Mappings
+            "original_genome_id <-> resolved_genome_id".
+        mappings_resolved_genome_id_bgc_id: Mappings
+            "resolved_genome_id <-> bgc_id".
+
+    Returns:
+        Key is strain id and value is a set of BGC ids.
+
+    See Also:
+        `extract_mappings_strain_id_original_genome_id`: Extract mappings
+            "strain_id <-> original_genome_id".
+        `extract_mappings_original_genome_id_resolved_genome_id`: Extract mappings
+            "original_genome_id <-> resolved_genome_id".
+        `extract_mappings_resolved_genome_id_bgc_id`: Extract mappings
+            "resolved_genome_id <-> bgc_id".
+    """
+    mappings_dict = {}
+    for strain_id, original_genome_ids in mappings_strain_id_original_genome_id.items():
+        bgc_ids = set()
+        for original_genome_id in original_genome_ids:
+            resolved_genome_id = mappings_original_genome_id_resolved_genome_id[original_genome_id]
+            if (bgc_id := mappings_resolved_genome_id_bgc_id.get(resolved_genome_id)) is not None:
+                bgc_ids.update(bgc_id)
+        if bgc_ids:
+            mappings_dict[strain_id] = bgc_ids
+    return mappings_dict
