+++ src/nplinker/metabolomics/__init__.py
@@ -1,9 +1,6 @@
 import logging
 from .molecular_family import MolecularFamily
 from .spectrum import Spectrum
-from .utils import add_annotation_to_spectrum
-from .utils import add_spectrum_to_mf
-from .utils import add_strains_to_spectrum
 
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
@@ -12,7 +9,4 @@ logging.getLogger(__name__).addHandler(logging.NullHandler())
 __all__ = [
     "MolecularFamily",
     "Spectrum",
-    "add_annotation_to_spectrum",
-    "add_spectrum_to_mf",
-    "add_strains_to_spectrum",
 ]
