--- src/nplinker/metabolomics/gnps/gnps_loader.py
@@ -1,3 +1,0 @@
-class GNPSLoader:
-    def __init__(self):
-        pass
