+++ src/nplinker/metabolomics/molecular_family.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from typing import TYPE_CHECKING
-from nplinker.strain import Strain
-from nplinker.strain_collection import StrainCollection
+from ..strain.strain import Strain
+from ..strain.strain_collection import StrainCollection
 
 
 if TYPE_CHECKING:
