+++ src/nplinker/metabolomics/utils.py
@@ -1,5 +1,10 @@
+import json
+from os import PathLike
+from pathlib import Path
 from nplinker.logconfig import LogConfig
-from nplinker.strain_collection import StrainCollection
+from nplinker.schemas import validate_podp_json
+from nplinker.strain import StrainCollection
+from .gnps.gnps_file_mapping_loader import GNPSFileMappingLoader
 from .molecular_family import MolecularFamily
 from .spectrum import Spectrum
 
@@ -113,3 +118,98 @@ def add_spectrum_to_mf(
         f"{len(mf_missing_spec)} MolecularFamily objects have missing Spectrum objects."
     )
     return mf_with_spec, mf_without_spec, mf_missing_spec
+
+
+# ------------------------------------------------------------------------------
+# Functions to extract mappings for metabolomics side:
+# strain_id <-> MS_filename <-> spectrum_id
+# ------------------------------------------------------------------------------
+def extract_mappings_strain_id_ms_filename(
+    podp_project_json_file: str | PathLike
+) -> dict[str, set[str]]:
+    """Extract mappings "strain_id <-> MS_filename".
+
+    Args:
+        podp_project_json_file: The path to the PODP project
+            JSON file.
+
+    Returns:
+        Key is strain id and value is a set of MS filenames.
+
+    Notes:
+        The `podp_project_json_file` is the project JSON file downloaded from
+        PODP platform. For example, for project MSV000079284, its json file is
+        https://pairedomicsdata.bioinformatics.nl/api/projects/4b29ddc3-26d0-40d7-80c5-44fb6631dbf9.4.
+    """
+    mappings_dict = {}
+    with open(podp_project_json_file, "r") as f:
+        json_data = json.load(f)
+
+    validate_podp_json(json_data)
+
+    # Extract mappings strain id <-> metabolomics filename
+    for record in json_data["genome_metabolome_links"]:
+        strain_id = record["genome_label"]
+        # get the actual filename of the mzXML URL
+        filename = Path(record["metabolomics_file"]).name
+        if strain_id in mappings_dict:
+            mappings_dict[strain_id].add(filename)
+        else:
+            mappings_dict[strain_id] = {filename}
+    return mappings_dict
+
+
+def extract_mappings_ms_filename_spectrum_id(
+    gnps_file_mappings_file: str | PathLike
+) -> dict[str, set[str]]:
+    """Extract mappings "MS_filename <-> spectrum_id".
+
+    Args:
+        gnps_file_mappings_file: The path to the GNPS file mappings file (csv or
+            tsv).
+
+    Returns:
+        Key is MS filename and value is a set of spectrum ids.
+
+    Notes:
+        The `gnps_file_mappings_file` is generated by GNPS molecular networking. It's downloaded
+        from GNPS website to a file with a default name defined in `GNPS_FILE_MAPPINGS_FILENAME`.
+
+    See Also:
+        GNPSFileMappingLoader: A class to load GNPS file mappings file.
+    """
+    loader = GNPSFileMappingLoader(gnps_file_mappings_file)
+    return loader.mapping_reversed
+
+
+def get_mappings_strain_id_spectrum_id(
+    mappings_strain_id_ms_filename: dict[str, set[str]],
+    mappings_ms_filename_spectrum_id: dict[str, set[str]],
+) -> dict[str, set[str]]:
+    """Get mappings "strain_id <-> spectrum_id".
+
+    Args:
+        mappings_strain_id_ms_filename: Mappings
+            "strain_id <-> MS_filename".
+        mappings_ms_filename_spectrum_id: Mappings
+            "MS_filename <-> spectrum_id".
+
+    Returns:
+        Key is strain id and value is a set of spectrum ids.
+
+
+    See Also:
+        `extract_mappings_strain_id_ms_filename`: Extract mappings
+            "strain_id <-> MS_filename".
+        `extract_mappings_ms_filename_spectrum_id`: Extract mappings
+            "MS_filename <-> spectrum_id".
+    """
+    mappings_dict = {}
+    for strain_id, ms_filenames in mappings_strain_id_ms_filename.items():
+        spectrum_ids = set()
+        for ms_filename in ms_filenames:
+            if (sid := mappings_ms_filename_spectrum_id.get(ms_filename)) is not None:
+                spectrum_ids.update(sid)
+        if spectrum_ids:
+            mappings_dict[strain_id] = spectrum_ids
+    return mappings_dict
