+++ src/nplinker/scoring/linking/data_links.py
@@ -13,7 +13,7 @@ from .utils import isinstance_all
 
 if TYPE_CHECKING:
     from nplinker.strain import Strain
-    from nplinker.strain_collection import StrainCollection
+    from nplinker.strain import StrainCollection
 
 logger = LogConfig.getLogger(__name__)
