+++ tests/genomics/test_mibig_loader.py
@@ -1,7 +1,7 @@
 import os.path
 import pytest
 from nplinker.genomics import BGC
-from nplinker.genomics import BGCLoaderBase
+from nplinker.genomics.abc import BGCLoaderBase
 from nplinker.genomics.mibig import MibigLoader
 from nplinker.genomics.mibig import download_and_extract_mibig_metadata
 from nplinker.genomics.mibig import parse_bgc_metadata_json
