+++ tests/genomics/test_podp_antismash_downloader.py
@@ -1,9 +1,9 @@
 import json
 from pathlib import Path
 import pytest
+from nplinker.genomics.antismash import GenomeStatus
+from nplinker.genomics.antismash import podp_download_and_extract_antismash_data
 from nplinker.globals import GENOME_STATUS_FILENAME
-from nplinker.strain import GenomeStatus
-from nplinker.strain import podp_download_and_extract_antismash_data
 from nplinker.utils import list_files
