+++ tests/metabolomics/test_spectrum.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 from nplinker.metabolomics import Spectrum
 from nplinker.strain import Strain
-from nplinker.strain_collection import StrainCollection
+from nplinker.strain import StrainCollection
 
 
 @pytest.mark.parametrize(
