+++ tests/scoring/conftest.py
@@ -7,7 +7,7 @@ from nplinker.scoring import MetcalfScoring
 from nplinker.scoring.linking import DataLinks
 from nplinker.scoring.linking import LinkFinder
 from nplinker.strain import Strain
-from nplinker.strain_collection import StrainCollection
+from nplinker.strain import StrainCollection
 
 
 @fixture(scope="session")
