--- tests/test_gnps_loader.py
@@ -1,6 +1,0 @@
-from nplinker.metabolomics.gnps.gnps_loader import GNPSLoader
-
-
-def test_default():
-    sut = GNPSLoader()
-    assert sut is not None
